mgr/cephadm: move SpecStore and HostCache to inventoy.py
Make module.py a bit smaller
